capnp Go : implement capnp imports and annotations.
